Update files according to suggestions.
